+++ src/config.h
@@ -15,6 +15,9 @@ class Config : public boost::noncopyable {
 public:
     virtual bool SetMaxBlockSize(uint64_t maxBlockSize) = 0;
     virtual uint64_t GetMaxBlockSize() const = 0;
+    virtual bool
+    SetBlockPriorityPercentage(int64_t blockPriorityPercentage) = 0;
+    virtual uint8_t GetBlockPriorityPercentage() const = 0;
     virtual const CChainParams &GetChainParams() const = 0;
 };
 
@@ -22,6 +25,8 @@ class GlobalConfig final : public Config {
 public:
     bool SetMaxBlockSize(uint64_t maxBlockSize);
     uint64_t GetMaxBlockSize() const;
+    bool SetBlockPriorityPercentage(int64_t blockPriorityPercentage);
+    uint8_t GetBlockPriorityPercentage() const;
     const CChainParams &GetChainParams() const;
 };