+++ src/globals.cpp
@@ -5,5 +5,7 @@
 #include "globals.h"
 
 #include "consensus/consensus.h"
+#include "policy/policy.h"
 
 uint64_t nMaxBlockSize = DEFAULT_MAX_BLOCK_SIZE;
+uint64_t nBlockPriorityPercentage = DEFAULT_BLOCK_PRIORITY_PERCENTAGE;